+++ CHANGES.rst
@@ -12,11 +12,25 @@ Changes
 
 Changes:
 --------
-- No change.
+- Add ``schema`` query parameter to ``GET /jobs/{jobID}/outputs`` request allowing to select between ``OGC``, ``OLD``
+  ``OGC+strict`` and ``OLD+strict`` representations (case insensitive), each with different combinations
+  of ``format.mimeType``, ``format.mediaType`` and/or directly ``type`` field to provide the Content-Type of an
+  output with ``href`` file.
+  By default, both the ``format`` (i.e.: ``OLD`` schema) and the ``type`` (i.e.: ``OGC`` schema) are simultaneously
+  reported for backward and forward compatibility, and for `OGC` compliance to return a file reference IANA Media-Type
+  (resolves `#401 <https://github.com/crim-ca/weaver/issues/401>`_).
+- Add support of ``type`` as alias to the Media-Type of the ``format`` of file references when submitted
+  for ``Job`` execution inputs, in accordance to the reported inputs/outputs endpoints, and for `OGC` compliance.
+- Drop ``type`` field for ``metadata`` items in process description that correspond to a ``value`` with a ``role``.
+- Enforce pattern validation of ``type`` as IANA Content-Type for ``metadata`` items in process description that
+  correspond to a ``Link`` with ``href``. Invalid ``type`` are now rejected to adhere to `OGC` requirement classes.
+- Clarify schema employed by `Weaver` to use naming that is as close as possible to `OGC` schemas to facilitate their
+  comprehension and external references.
 
 Fixes:
 ------
-- No change.
+- Fix ``GET /jobs/{jobID}/inputs`` endpoint failing to return submitted ``inputs`` when they were specified using the
+  mapping representation (i.e.: ``OGC`` schema).
 
 .. _changes_4.12.0: